backend/requirements.txt
alembic
asyncpg
# Core Web Framework
fastapi

# HTTP Client
httpx

# AI Stack (Latest Compatible Versions)
langchain
langchain-openai
langgraph
langsmith
openai

# Monitoring
prometheus-client
pydantic
pydantic-settings

# Document Processing
pypdf

# Image Processing
Pillow

# Testing
pytest
pytest-asyncio
pytest-cov

# Environment & Config
python-dotenv

# Utils
pyyaml

# Vector Store
qdrant-client

# Cache
redis[asyncio]>=5.0.0

# Text Processing
sentence-transformers

# Epic 3 - Pattern Retrieval
rank-bm25>=0.2.2
tenacity>=8.2.0

# Database
sqlalchemy
psycopg2-binary
greenlet

# Server-Sent Events
sse-starlette

# Epic 6 - Production Infrastructure
boto3>=1.28.0

# Epic 8 - Regeneration
celery[redis]>=5.3.0

# Epic 3 - Safety & Guardrails
nh3>=0.2.0
python-magic>=0.4.27
slowapi>=0.1.9

# Document Processing
unstructured[pdf]
uvicorn[standard]

# Jupyter and Data Science
jupyter
ipykernel
matplotlib
pandas
scipy
seaborn
plotly
notebook
nest-asyncio
